Move investigation code to dedicated module
