+++ backend/src/investigation/structs.rs
@@ -1,0 +1,38 @@
+use axum::{
+    Json,
+    response::{IntoResponse, Response},
+};
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, ToSchema)]
+#[serde(deny_unknown_fields)]
+pub struct PollingStationInvestigation {
+    pub polling_station_id: u32,
+    pub reason: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[schema(nullable = false)]
+    pub findings: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[schema(nullable = false)]
+    pub corrected_results: Option<bool>,
+}
+
+impl IntoResponse for PollingStationInvestigation {
+    fn into_response(self) -> Response {
+        Json(self).into_response()
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, ToSchema)]
+#[serde(deny_unknown_fields)]
+pub struct PollingStationInvestigationCreateRequest {
+    pub reason: String,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, ToSchema)]
+#[serde(deny_unknown_fields)]
+pub struct PollingStationInvestigationConcludeRequest {
+    pub findings: String,
+    pub corrected_results: bool,
+}
